requirement.txt
langchain
langchain-core
langchain_community

langchain-openai
openai

langchain-anthropic

langchain-google-genai
google-generativeai

langchain-huggingface
transformers
huggingface-hub


python-dotenv

numpy 
scikit-learn